Bucket o' changes to get to a simple path to score minimum number of points in first competition.
LinearSlide not functioning so changed that functionality.
simplified drive path to just score 10 points based on recognizing the standard signal sleeve.

simplified run path to only move to the parking space to score points.

made changes to LinearSlide constants and SwingArm to attempt to just move to positions - made target position to class level to only move to positions.